requirements.txt
Pillow==8.0.1
streamlit==0.69.2
tensorflow==2.3.1
gunicorn==20.0.4

